Fix date range modal to overlay full app including sidebar
The CustomDateRangeModal was embedded inside DashboardPage and
AnalyticsPage, causing it to only cover the page content area
(not the sidebar) and take up the full page height. Moved the
modal to the AppShell root Panel level, matching the pattern
used by all other modals in the app.

- Created CustomDateRangeModalViewModel with callback-based Open()
  method shared between Dashboard and Analytics pages
- Updated CustomDateRangeModal.axaml to follow ConfirmationDialog
  pattern (IsVisible, centering, ClosingCommand, x:DataType)
- Registered VM in AppShellViewModel and App.axaml.cs
- Removed duplicated modal state from page ViewModels
