+++ ArgoBooks/ViewModels/AnalyticsPageViewModel.cs
@@ -175,19 +175,6 @@ public partial class AnalyticsPageViewModel : ChartContextMenuViewModelBase
         }
     }
 
-    // Temporary date values for the modal (before applying)
-    [ObservableProperty]
-    private DateTime _modalStartDate = new(DateTime.Now.Year, DateTime.Now.Month, 1);
-
-    [ObservableProperty]
-    private DateTime _modalEndDate = DateTime.Now;
-
-    /// <summary>
-    /// Gets or sets whether the custom date range modal is open.
-    /// </summary>
-    [ObservableProperty]
-    private bool _isCustomDateRangeModalOpen;
-
     /// <summary>
     /// Gets or sets whether a custom date range has been applied (delegates to shared service).
     /// </summary>
@@ -243,36 +230,6 @@ public partial class AnalyticsPageViewModel : ChartContextMenuViewModelBase
     }
 
     /// <summary>
-    /// Gets or sets the modal start date as DateTimeOffset for DatePicker binding.
-    /// </summary>
-    public DateTimeOffset? ModalStartDateOffset
-    {
-        get => new DateTimeOffset(ModalStartDate);
-        set
-        {
-            if (value.HasValue)
-            {
-                ModalStartDate = value.Value.DateTime;
-            }
-        }
-    }
-
-    /// <summary>
-    /// Gets or sets the modal end date as DateTimeOffset for DatePicker binding.
-    /// </summary>
-    public DateTimeOffset? ModalEndDateOffset
-    {
-        get => new DateTimeOffset(ModalEndDate);
-        set
-        {
-            if (value.HasValue)
-            {
-                ModalEndDate = value.Value.DateTime;
-            }
-        }
-    }
-
-    /// <summary>
     /// Gets whether the custom date range option is selected.
     /// </summary>
     public bool IsCustomDateRange => SelectedDateRange == "Custom Range";
@@ -285,7 +242,6 @@ public partial class AnalyticsPageViewModel : ChartContextMenuViewModelBase
     /// <summary>
     /// Opens the custom date range modal.
     /// </summary>
-    [RelayCommand]
     private void OpenCustomDateRangeModal()
     {
         // Store the previous selection before opening the modal
@@ -295,67 +251,24 @@ public partial class AnalyticsPageViewModel : ChartContextMenuViewModelBase
             _previousDateRange = SelectedDateRange;
         }
 
-        // Initialize modal dates with current values
-        ModalStartDate = StartDate;
-        ModalEndDate = EndDate;
-        OnPropertyChanged(nameof(ModalStartDateOffset));
-        OnPropertyChanged(nameof(ModalEndDateOffset));
-        IsCustomDateRangeModalOpen = true;
-    }
-
-    /// <summary>
-    /// Applies the custom date range from the modal.
-    /// </summary>
-    [RelayCommand]
-    private async Task ApplyCustomDateRange()
-    {
-        // Check if start date is after end date
-        if (ModalStartDate > ModalEndDate)
-        {
-            var result = await App.ConfirmationDialog!.ShowAsync(new ConfirmationDialogOptions
+        App.CustomDateRangeModal?.Open(StartDate, EndDate,
+            onApply: (start, end) =>
             {
-                Title = "Invalid Date Range",
-                Message = "The start date is after the end date. Would you like to swap the dates?",
-                PrimaryButtonText = "Swap Dates",
-                CancelButtonText = "Cancel"
+                StartDate = start;
+                EndDate = end;
+                HasAppliedCustomRange = true;
+                OnPropertyChanged(nameof(AppliedDateRangeText));
+                LoadAllCharts();
+            },
+            onCancel: () =>
+            {
+                // If no custom range was previously applied, revert to the previous selection
+                if (!HasAppliedCustomRange)
+                {
+                    ChartSettingsShared.SelectedDateRange = _previousDateRange;
+                    OnPropertyChanged(nameof(SelectedDateRange));
+                }
             });
-
-            if (result == ConfirmationResult.Primary)
-            {
-                // Swap the dates
-                (ModalStartDate, ModalEndDate) = (ModalEndDate, ModalStartDate);
-                OnPropertyChanged(nameof(ModalStartDateOffset));
-                OnPropertyChanged(nameof(ModalEndDateOffset));
-            }
-            else
-            {
-                // User cancelled, keep modal open
-                return;
-            }
-        }
-
-        StartDate = ModalStartDate;
-        EndDate = ModalEndDate;
-        HasAppliedCustomRange = true;
-        OnPropertyChanged(nameof(AppliedDateRangeText));
-        IsCustomDateRangeModalOpen = false;
-        LoadAllCharts();
-    }
-
-    /// <summary>
-    /// Cancels the custom date range modal.
-    /// </summary>
-    [RelayCommand]
-    private void CancelCustomDateRange()
-    {
-        IsCustomDateRangeModalOpen = false;
-
-        // If no custom range was previously applied, revert to the previous selection
-        if (!HasAppliedCustomRange)
-        {
-            ChartSettingsShared.SelectedDateRange = _previousDateRange;
-            OnPropertyChanged(nameof(SelectedDateRange));
-        }
     }
 
     #endregion
