+++ ArgoBooks/Views/DashboardPage.axaml
@@ -4,7 +4,6 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:vm="using:ArgoBooks.ViewModels"
              xmlns:controls="using:ArgoBooks.Controls"
-             xmlns:modals="using:ArgoBooks.Modals"
              xmlns:local="using:ArgoBooks"
              xmlns:lvc="using:LiveChartsCore.SkiaSharpView.Avalonia"
              xmlns:loc="using:ArgoBooks.Localization"
@@ -786,7 +785,5 @@
                                ResetChartZoomCommand="{Binding ResetChartZoomCommand}"
                                CloseCommand="{Binding HideChartContextMenuCommand}" />
 
-    <!-- Custom Date Range Modal -->
-    <modals:CustomDateRangeModal />
     </Panel>
 </UserControl>
